chore: hoist calculator dash components to root
