+++ app/survey/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from 'react';
 import { Inter } from "next/font/google";
-import ErrorBoundary from '../components/ErrorBoundary';
+import ErrorBoundary from '../../components/custom/ErrorBoundary';
 import { SurveyResults } from './types';
 import { Age } from './components/graphs/Age';
 import { AffordFairhold } from './components/graphs/AffordFairhold';
