+++ components/custom/dashboard/graphs/CostOverTimeStackedBarChart.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Bar, BarChart, CartesianGrid, XAxis, YAxis, Tooltip } from "recharts";
-import { Card, CardContent } from "@/components/ui/card";
+import { Card, CardContent } from "@/components/card";
 import {
   StyledChartContainer,
 } from "../ui/StyledChartContainer";
