+++ components/custom/dashboard/graphs/CostOverTimeStackedBarChartMobile.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Bar, BarChart, CartesianGrid, XAxis, YAxis, Tooltip } from "recharts";
-import { Card, CardContent } from "@/components/ui/card";
+import { Card, CardContent } from "@/components/card";
 import { StyledChartContainer } from "../ui/StyledChartContainer";
 import { formatValue } from "@/app/lib/format";
 import { LifetimeBarData } from "./CostOverTimeStackedBarChart";
