+++ components/custom/dashboard/graphs/CostOverTimeWrapper.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import ErrorBoundary from "../ErrorBoundary";
+import ErrorBoundary from "../../ErrorBoundary";
 import { Household } from "@/app/models/Household";
 import CostOverTimeStackedBarChart, { LifetimeBarData } from "./CostOverTimeStackedBarChart";
 import CostOverTimeStackedBarChartMobile from "./CostOverTimeStackedBarChartMobile";
