+++ components/custom/ui/GraphCard.tsx
