+++ components/custom/ui/InputDropdown.tsx
@@ -5,7 +5,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select";
+} from "@/components/select";
 
 interface DropdownOption {
   readonly label: string;
