+++ components/custom/ui/InputField.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Input } from "@/components/ui/input";
+import { Input } from "@/components/input";
 
 interface InputFieldProps extends React.InputHTMLAttributes<HTMLInputElement> {
   placeholder: string;
